playground/python/pytest-playwright-combo-python/requirements.txt
pytest
playwright
appium-python-client
requests
python-dotenv

pytest-rerunfailures